requirements.txt
# FastAPI 核心依赖
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# 数据库
sqlalchemy>=2.0.0

# 配置管理
pydantic>=2.5.0
pydantic-settings>=2.1.0

# 模板引擎
jinja2>=3.1.2

# 数据处理（分析脚本用）
pandas>=2.1.0
numpy>=1.26.0
openpyxl>=3.1.0

# 图片处理（可选）
pillow>=10.0.0

# 定时任务
apscheduler>=3.10.0

# 生产服务器（可选）
# hypercorn>=0.15.0

# 测试依赖
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0
pytest-cov>=4.1.0  # 覆盖率报告
